Removed telegram messenger as I don't plan to test/use it. Creating cointegrated pairs is working now.

+++ program/func_cointegration.py
@@ -32,8 +32,8 @@ def calculate_zscore(spread):
 
 # Calculate Cointegration
 def calculate_cointegration(series_1, series_2):
-  series_1 = np.array(series_1).astype(np.float)
-  series_2 = np.array(series_2).astype(np.float)
+  series_1 = np.array(series_1).astype(np.float32) # Pandas doesn't have "float" datatype any longer. "float32" should be sufficient precise for price data.
+  series_2 = np.array(series_2).astype(np.float32)
   coint_flag = 0
   coint_res = coint(series_1, series_2)
   coint_t = coint_res[0]
